requirements.txt
# Agentic Hub Backend Dependencies
openai>=1.3.0
Jinja2
mlflow==3.8.0
PyYAML
pgvector>=0.2.5
psycopg[binary,pool]>=3.1.19
jsonschema>=4.22.0
requests
PyMuPDF
camelot-py

# Web Framework
fastapi
uvicorn

# Data Validation
pydantic>=2.5.0

# CORS
python-multipart>=0.0.6

## Utils
python-dotenv
# For token counting in LLM interactions
tiktoken>=0.5.1

# TODO: Add when implementing actual persistence
# sqlalchemy>=2.0.0
# asyncpg>=0.29.0  # PostgreSQL async driver
# redis>=5.0.0     # For caching

# TODO: Add when implementing MLflow integration
# mlflow>=2.10.0

# TODO: Add when implementing observability
# opentelemetry-api>=1.22.0
# opentelemetry-sdk>=1.22.0
# opentelemetry-exporter-jaeger>=1.22.0
